Simplify: extract useLandscapeScreenDimensions from player (#16 slice 6)
Move the landscape-fullscreen screen-dimension tracking (two Reanimated shared
values + the Dimensions 'change' listener effect) out of the 3,023-line
VideoPlayerOverlayImpl into lib/useLandscapeScreenDimensions.ts. The block was
fully self-contained — the shared values are written only by this effect and
read only in two animated styles — so it's pure code motion: the component
destructures { landscapeWidth, landscapeHeight } from the hook and the animated
styles are unchanged. Hook is called unconditionally at the same position, so
hook order/effect timing is preserved.

The remaining player bulk (mobile drag-gesture physics, fullscreen/landscape
animation, PiP geometry) stays inlined — it's gesture/animation-coupled and tsc
can't validate behavior, so those slices belong in a run-capable environment.

Validated: app tsc shows zero new errors (the one shifted line is the
pre-existing StyleSheet.absoluteFillObject error, relocated by the deletion).
